mcp-server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.34.0
sqlmodel==0.0.22
sqlalchemy[asyncpg]==2.0.36
pydantic==2.10.3
python-dotenv==1.0.1
mcp>=1.0.0
pytest==8.3.4
pytest-asyncio==0.23.0
httpx==0.28.0
